refactor(level-up): split getSubclassFeatures into focused single-export modules
Extract types to types/subclassFeatures.d.ts, move builder to its own
file, and simplify the lookup to a concise single-export module.
Deduplicate world-items/packs processing and parallelize pack indexing.
