+++ src/view/dialogs/CharacterLevelUpDialog.test.ts
@@ -17,8 +17,10 @@ vi.mock('../../utils/getClassFeatures.ts', () => ({
 // Mock the subclass features utilities to return empty results immediately.
 // We mock via the path-alias that the source file uses so both the test and the
 // source resolve to the same mock instance, enabling vi.spyOn on the module.
+vi.mock('#utils/buildSubclassFeatureIndex.ts', () => ({
+	default: vi.fn(() => Promise.resolve(new Map())),
+}));
 vi.mock('#utils/getSubclassFeatures.ts', () => ({
-	buildSubclassFeatureIndex: vi.fn(() => Promise.resolve(new Map())),
 	default: vi.fn(() => Promise.resolve([])),
 }));
